Add s3_compatible storage option for all S3 compatible storage targets
Further consolidation of the code, s3_compatible is the new s3 compatible storage driver
Fix some bugs, typos and make the s3_compat load credentials with the aws keys (for compatibility)
Address comments, add http/https differentation to endpoint_url generation
Remove is_azure from Azure blobs storage as backup_node does not need it
(if threshold is set, it will be used)
(cherry picked from commit 544940dda0463df7e062113ad498345fa97a5955)
Remove IBM storage implementation, use S3BaseStorage for ibm_storage
Use Provider.S3 as storage driver always in the AWS S3 storage
Add backwards compatibility for the old s3 driver name format, modify one test to use the old way
Fix IBM Cloud Storage test setup.
