+++ medusa/backup_node.py
@@ -30,8 +30,6 @@ import medusa.utils
 from medusa.cassandra_utils import Cassandra
 from medusa.index import add_backup_start_to_index, add_backup_finish_to_index, set_latest_backup_in_index
 from medusa.monitoring import Monitoring
-from medusa.storage.s3_storage import is_aws_s3
-from medusa.storage.azure_storage import is_azure
 from medusa.storage.google_storage import GSUTIL_MAX_FILES_PER_CHUNK
 from medusa.storage import Storage, format_bytes_str, ManifestObject, divide_chunks
 
@@ -86,8 +84,7 @@ class NodeBackupCache(object):
                 if self._storage_provider == Provider.GOOGLE_STORAGE or self._differential_mode is True:
                     cached_item = self._cached_objects.get(fqtn, {}).get(src.name)
 
-                threshold = self._storage_config.multi_part_upload_threshold \
-                    if is_aws_s3(self._storage_provider) or is_azure(self._storage_provider) else None
+                threshold = self._storage_config.multi_part_upload_threshold
                 if cached_item is None or not self._storage_driver.file_matches_cache(src, cached_item, threshold):
                     # We have no matching object in the cache matching the file
                     retained.append(src)
