+++ medusa/storage/__init__.py
@@ -21,7 +21,7 @@ import operator
 import pathlib
 import re
 
-from libcloud.storage.providers import Provider
+from libcloud.storage.providers import Provider, get_driver
 from libcloud.common.types import InvalidCredsError
 from retrying import retry
 
@@ -34,7 +34,7 @@ from medusa.storage.local_storage import LocalStorage
 from medusa.storage.s3_storage import S3Storage
 from medusa.storage.s3_rgw import S3RGWStorage
 from medusa.storage.azure_storage import AzureStorage
-from medusa.storage.ibm_storage import IBMCloudStorage
+from medusa.storage.s3_base_storage import S3BaseStorage
 
 
 ManifestObject = collections.namedtuple('ManifestObject', ['path', 'size', 'MD5'])
@@ -73,6 +73,7 @@ class Storage(object):
         self.storage_provider = self._config.storage_provider
 
     def _connect_storage(self):
+        logging.debug('Loading storage_provider: {}'.format(self._config.storage_provider))
         if self._config.storage_provider == Provider.GOOGLE_STORAGE:
             google_storage = GoogleStorage(self._config)
             google_storage.check_dependencies()
@@ -83,14 +84,23 @@ class Storage(object):
             return azure_storage
         elif self._config.storage_provider == Provider.S3_RGW:
             return S3RGWStorage(self._config)
+        elif self._config.storage_provider.lower() == "s3_compatible":
+            s3_storage = S3BaseStorage(self._config)
+            s3_storage.check_dependencies()
+            return s3_storage
         elif self._config.storage_provider.startswith(Provider.S3):
+            if self._config.storage_provider != Provider.S3:
+                self._config['region'] = get_driver(self._config.storage_provider).region_name
+
             s3_storage = S3Storage(self._config)
             s3_storage.check_dependencies()
             return s3_storage
         elif self._config.storage_provider == Provider.LOCAL:
             return LocalStorage(self._config)
         elif self._config.storage_provider.lower() == "ibm_storage":
-            return IBMCloudStorage(self._config)
+            s3_storage = S3BaseStorage(self._config)
+            s3_storage.check_dependencies()
+            return s3_storage
 
         raise NotImplementedError("Unsupported storage provider")
